redis test app updates
