+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "async-redis": "^2.0.0"
+  }
+}
