+++ test/magpie/package.json
@@ -22,6 +22,7 @@
     "@azure/keyvault-secrets": "^4.2.0",
     "@azure/service-bus": "^7.3.0",
     "@roadwork/dapr-js-sdk": "^0.3.1",
+    "bluebird": "^3.7.2",
     "express": "^4.17.1",
     "mongodb": "^3.6.10",
     "redis": "^3.1.2"
